urine_strip_analyzer/requirements.txt
Django
streamlit
opencv-python-headless
Pillow
numpy
requests
